Rename (stripped) SimpleNuGet to Sharpmake.NuGet
