+++ Sharpmake.NuGet/INuGetDependency.cs
@@ -11,25 +11,23 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-using System;
-using System.Xml.Linq;
+using System.Collections.Generic;
 
-namespace SimpleNuGet.Impl
+namespace Sharpmake.NuGet
 {
-    internal static class XElementExtensions
+    /// <summary>
+    /// Dependency between packages
+    /// </summary>
+    public interface INuGetDependency
     {
-        public static string GetOptionalAttributeValue(this XElement element, string localName, string namespaceName = null)
-        {
-            XAttribute attr;
-            if (String.IsNullOrEmpty(namespaceName))
-            {
-                attr = element.Attribute(localName);
-            }
-            else
-            {
-                attr = element.Attribute(XName.Get(localName, namespaceName));
-            }
-            return attr != null ? attr.Value : null;
-        }
+        /// <summary>
+        /// The package ID of the dependency.
+        /// </summary>
+        string Name { get; }
+
+        /// <summary>
+        /// The range of versions acceptable as a dependency. See Dependency versions for exact syntax.
+        /// </summary>
+        VersionRange VersionRange { get; }
     }
 }
