+++ Sharpmake.NuGet/Impl/JsonObjectWrapper.cs
@@ -14,7 +14,7 @@
 
 using System.IO;
 
-namespace SimpleNuGet.Impl
+namespace Sharpmake.NuGet.Impl
 {
     /// <summary>
     /// Provides a typed array interface around a json object where each key represent an item in the list.
