+++ Sharpmake.NuGet/Properties/AssemblyInfo.cs
@@ -18,11 +18,11 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("SimpleNuGet")]
+[assembly: AssemblyTitle("Sharpmake.NuGet")]
 [assembly: AssemblyDescription("")]
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("Ubisoft")]
-[assembly: AssemblyProduct("SimpleNuGet")]
+[assembly: AssemblyProduct("Sharpmake.NuGet")]
 [assembly: AssemblyCopyright("Copyright © Ubisoft 2017")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
@@ -47,5 +47,3 @@ using System.Runtime.InteropServices;
 // [assembly: AssemblyVersion("1.0.*")]
 [assembly: AssemblyVersion("1.0.0.0")]
 [assembly: AssemblyFileVersion("1.0.0.0")]
-
-[assembly: InternalsVisibleTo("SimpleNuGetTests")]
