+++ Sharpmake/PackageReferences.cs
@@ -16,7 +16,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Text.RegularExpressions;
-using SimpleNuGet;
+using Sharpmake.NuGet;
 
 namespace Sharpmake
 {
