+++ deploy_binaries.py
@@ -108,7 +108,7 @@ copy_list = [
     BinarySite("Sharpmake", "Sharpmake"),
     BinarySite("Sharpmake.Application", "Sharpmake.Application"),
     BinarySite("Sharpmake.Generators", "Sharpmake.Generators"),
-    BinarySite("SimpleNuGet", "SimpleNuGet")
+    BinarySite("Sharpmake.NuGet", "Sharpmake.NuGet")
 ]
 
 # Add the platforms to the list of files to copy.
