+++ samples/SharpmakeGen/reference/Sharpmake.sln
@@ -26,6 +26,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sharpmake.Durango", "Sharpm
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sharpmake.Generators", "Sharpmake.Generators\Sharpmake.Generators.csproj", "{844F66DE-B015-340E-720A-8E158B517E93}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sharpmake.NuGet", "Sharpmake.NuGet\Sharpmake.NuGet.csproj", "{4DDFA6FA-D217-3CF5-0E65-BB401B9D4A01}"
+EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sharpmake.NvShield", "Sharpmake.Platforms\Sharpmake.NvShield\Sharpmake.NvShield.csproj", "{C05B1136-5243-BFF8-F39B-24AB8261FD54}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Sharpmake.UnitTests", "Sharpmake.UnitTests\Sharpmake.UnitTests.csproj", "{BE99068F-E9D0-EF2F-1ECE-CAAA3D6E2DF0}"
@@ -35,8 +37,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharpmakeGen", "samples\SharpmakeGen\SharpmakeGen.csproj", "{C2F8D8B8-29F8-4117-62B6-F6E3E63C225B}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleExeLibDependency", "samples\SimpleExeLibDependency\SimpleExeLibDependency.csproj", "{128E20D6-0408-DD6C-4BA5-0B7A051E6734}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SimpleNuGet", "SimpleNuGet\SimpleNuGet.csproj", "{D163B9B5-0A3D-9F09-F47E-83EB1485E2C0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -88,6 +88,10 @@ Global
 		{844F66DE-B015-340E-720A-8E158B517E93}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{844F66DE-B015-340E-720A-8E158B517E93}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{844F66DE-B015-340E-720A-8E158B517E93}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4DDFA6FA-D217-3CF5-0E65-BB401B9D4A01}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4DDFA6FA-D217-3CF5-0E65-BB401B9D4A01}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4DDFA6FA-D217-3CF5-0E65-BB401B9D4A01}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4DDFA6FA-D217-3CF5-0E65-BB401B9D4A01}.Release|Any CPU.Build.0 = Release|Any CPU
 		{C05B1136-5243-BFF8-F39B-24AB8261FD54}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{C05B1136-5243-BFF8-F39B-24AB8261FD54}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{C05B1136-5243-BFF8-F39B-24AB8261FD54}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -108,10 +112,6 @@ Global
 		{128E20D6-0408-DD6C-4BA5-0B7A051E6734}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{128E20D6-0408-DD6C-4BA5-0B7A051E6734}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{128E20D6-0408-DD6C-4BA5-0B7A051E6734}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D163B9B5-0A3D-9F09-F47E-83EB1485E2C0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D163B9B5-0A3D-9F09-F47E-83EB1485E2C0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D163B9B5-0A3D-9F09-F47E-83EB1485E2C0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D163B9B5-0A3D-9F09-F47E-83EB1485E2C0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
